Move wizard validation into fermenter
