+++ tests/test_wizard_value_list_integration.py
@@ -5,10 +5,7 @@ from pathlib import Path
 import streamlit as st
 
 import gkc
-from gkc.profiles.forms.validation_bridge import (
-    _merge_wikibase_item_metadata,
-    validate_inline_value,
-)
+from gkc.fermenter import validate_inline_value
 from gkc.wizard.steps import (
     _coerce_nested_statement_map,
     _extract_value_list_candidates,
@@ -64,26 +61,6 @@ def test_filter_value_list_candidates_matches_label_qid_and_uri() -> None:
     assert (
         len(_filter_value_list_candidates(candidates, "wikidata.org/entity/q42")) == 1
     )
-
-
-def test_merge_wikibase_item_metadata_preserves_uri_and_label() -> None:
-    original = {
-        "id": "Q42",
-        "item": "http://www.wikidata.org/entity/Q42",
-        "itemLabel": "Douglas Adams",
-    }
-    normalized = {
-        "entity-type": "item",
-        "numeric-id": 42,
-        "id": "Q42",
-    }
-
-    merged = _merge_wikibase_item_metadata(original, normalized)
-
-    assert merged["id"] == "Q42"
-    assert merged["item"] == "http://www.wikidata.org/entity/Q42"
-    assert merged["itemLabel"] == "Douglas Adams"
-
 
 def test_validate_inline_value_keeps_item_metadata() -> None:
     value = {
